- removing archetype query stuff, not needed currently - cleaning up a bit
